requirements.txt
Pillow
Pillow==8.1.0
pytube==10.4.1
typing-extensions==3.7.4.3
